simplify CallHandler and Handler etc
