--- lib/delfos/method_trace/return_handler.rb
@@ -1,20 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "handler"
-require "delfos/call_stack"
-
-module Delfos
-  module MethodTrace
-    class ReturnHandler < Handler
-      def perform
-        return unless relevant?
-
-        begin
-          CallStack.pop
-        rescue
-          nil
-        end
-      end
-    end
-  end
-end
